Extracted utility functions to utils.ts Created tests for the tax breakdown calculation behaviour
