+++ .env.test
@@ -1,3 +1,4 @@
 VITE_RC_ENDPOINT=http://localhost:8000
 VITE_RC_ANALYTICS_ENDPOINT=http://localhost:8000
 VITE_ASSETS_ENDPOINT=https://da08ctfrofx1b.cloudfront.net
+VITE_ALLOW_TAX_CALCULATION_FF=true
